problem(203): remove linked list elements
