+++ src/141-(easy)-linked_list_cycle.ts
@@ -1,17 +1,5 @@
 // link to problem: https://leetcode.com/problems/linked-list-cycle
-
-/**
- * Definition for singly-linked list.
- *
- */
-class ListNode {
-  val: number
-  next: ListNode | null
-  constructor(val?: number, next?: ListNode | null) {
-    this.val = (val === undefined ? 0 : val)
-    this.next = (next === undefined ? null : next)
-  }
-}
+import { ListNode } from "./helpers/list-node";
 
 function hasCycle(head: ListNode | null): boolean {
   const newCollection = new Set<ListNode>();
