+++ src/helpers/list-node.ts
@@ -1,0 +1,12 @@
+/**
+ * Definition for singly-linked list.
+ *
+ */
+export class ListNode {
+  val: number
+  next: ListNode | null
+  constructor(val?: number, next?: ListNode | null) {
+    this.val = (val === undefined ? 0 : val)
+    this.next = (next === undefined ? null : next)
+  }
+}
